backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
SQLAlchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1
python-dotenv==1.0.1
pandas